+++ CHANGELOG.md
@@ -9,9 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add debug mode option
+- Add paste and match style to edit menu
+
 ### Changed
 
-### Fixed
+- Updated to Typescript
 
 ## [2.2.0]